Выносит логику отрисовки фильмов в контроллер.WIP
